chore(instrument): rework test files
Use `instrument_test` package for tests of the public API. This ensures
all test packaging styles are in use in this package; which is a good
thing for some package-loading tests.
