+++ instrument/instrument.go
@@ -14,33 +14,6 @@ import (
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
 )
 
-func getOpName(metadata ...any) string {
-	rank := map[string]int{
-		"verb":          1,
-		"function-name": 2,
-	}
-
-	var (
-		opname string
-		oprank = 10_000 // just a higher number than any key in the rank map.
-	)
-	for i := 0; i < len(metadata); i += 2 {
-		if i+1 >= len(metadata) {
-			break
-		}
-		if k, ok := metadata[i].(string); ok {
-			if r, ok := rank[k]; ok && r < oprank {
-				if on, ok := metadata[i+1].(string); ok {
-					opname = on
-					oprank = r
-					continue
-				}
-			}
-		}
-	}
-	return opname
-}
-
 func Report(ctx context.Context, e event.Event, metadata ...any) context.Context {
 	var span tracer.Span
 	if e == event.EventStart || e == event.EventCall {
