+++ instrument/instrument_test.go
@@ -3,12 +3,13 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2023-present Datadog, Inc.
 
-package instrument
+package instrument_test
 
 import (
 	"context"
 	"testing"
 
+	"github.com/DataDog/orchestrion/instrument"
 	"github.com/DataDog/orchestrion/instrument/event"
 
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
@@ -17,7 +18,7 @@ import (
 func TestReport(t *testing.T) {
 	t.Run("start", func(t *testing.T) {
 		ctx := context.Background()
-		ctx = Report(ctx, event.EventStart)
+		ctx = instrument.Report(ctx, event.EventStart)
 		if _, ok := tracer.SpanFromContext(ctx); !ok {
 			t.Errorf("Expected Report of StartEvent to generate a new ID.")
 		}
@@ -25,7 +26,7 @@ func TestReport(t *testing.T) {
 
 	t.Run("call", func(t *testing.T) {
 		ctx := context.Background()
-		ctx = Report(ctx, event.EventCall)
+		ctx = instrument.Report(ctx, event.EventCall)
 		if _, ok := tracer.SpanFromContext(ctx); !ok {
 			t.Errorf("Expected Report of CallEvent to generate a new ID.")
 		}
@@ -33,7 +34,7 @@ func TestReport(t *testing.T) {
 
 	t.Run("end", func(t *testing.T) {
 		ctx := context.Background()
-		ctx = Report(ctx, event.EventEnd)
+		ctx = instrument.Report(ctx, event.EventEnd)
 		if _, ok := tracer.SpanFromContext(ctx); ok {
 			t.Errorf("Expected Report of EndEvent not to generate a new ID.")
 		}
@@ -41,41 +42,9 @@ func TestReport(t *testing.T) {
 
 	t.Run("return", func(t *testing.T) {
 		ctx := context.Background()
-		ctx = Report(ctx, event.EventReturn)
+		ctx = instrument.Report(ctx, event.EventReturn)
 		if _, ok := tracer.SpanFromContext(ctx); ok {
 			t.Errorf("Expected Report of ReturnEvent not to generate a new ID.")
 		}
 	})
-}
-
-func TestGetOpName(t *testing.T) {
-	for _, tt := range []struct {
-		metadata []any
-		opname   string
-	}{
-		{
-			metadata: []any{"foo", "bar", "verb", "just-verb"},
-			opname:   "just-verb",
-		},
-		{
-			metadata: []any{"foo", "bar", "function-name", "just-function-name"},
-			opname:   "just-function-name",
-		},
-		{
-			metadata: []any{"foo", "bar", "verb", "verb-function-name", "function-name", "THIS IS WRONG"},
-			opname:   "verb-function-name",
-		},
-		{
-			// Checking different order
-			metadata: []any{"foo", "bar", "function-name", "THIS IS WRONG", "verb", "verb-function-name"},
-			opname:   "verb-function-name",
-		},
-	} {
-		t.Run(tt.opname, func(t *testing.T) {
-			n := getOpName(tt.metadata...)
-			if n != tt.opname {
-				t.Errorf("Expected %s, but got %s\n", tt.opname, n)
-			}
-		})
-	}
 }
